init refactoring
