+++ assets/styles/_font.scss
@@ -1,69 +1,7 @@
 @font-face {
-    font-family: 'Roboto Mono';
-    src: url("~assets/fonts/RobotoMono.ttf") format('truetype');
-}
-
-@font-face {
-    font-family: 'Roboto Mono';
-    font-weight: 500;
-    src: url("~assets/fonts/RobotoMono-Medium.ttf") format('truetype');
-}
-
-@font-face {
-    font-family: 'Roboto';
-    font-style: normal;
-    font-weight: 700;
-    src: url("~assets/fonts/Roboto-Bold.ttf") format('truetype');
-}
-
-@font-face {
-    font-family: 'Roboto';
-    font-style: normal;
-    font-weight: 400;
-    src: url("~assets/fonts/Roboto-Regular.ttf") format('truetype');
-}
-
-@font-face {
-    font-family: 'Roboto';
-    font-weight: 500;
-    src: url("~assets/fonts/Roboto-Medium.ttf") format('truetype');
-}
-
-// @font-face {
-//     font-family: 'Roboto';
-//     src: url("~assets/fonts/Roboto-Regular.ttf") format('truetype');
-// }
-@font-face {
-    font-family: 'Exo 2';
-    font-weight: 100 1000;
-    font-stretch: 25% 151%;
-    src: url("~assets/fonts/Exo2.ttf") format('truetype');
-}
-@font-face {
-    font-family: 'Exo 2';
-    font-weight: 100 1000;
-    font-stretch: 25% 151%;
-    font-style: italic;
-    src: url("~assets/fonts/Exo_2/Exo2-Italic-VariableFont_wght.ttf") format('truetype');
-}
-
-@font-face {
     font-family: 'ProductSans';
     font-weight: 100 1000;
     font-stretch: 25% 151%;
     font-style: normal;
     src: url("~assets/fonts/Product Sans Regular.ttf") format('truetype');
 }
-// @font-face {
-//     font-family: 'Exo 2';
-//     font-style: normal;
-//     font-weight: 700;
-//     src: url("~assets/fonts/Exo_2/static/Exo2-Bold.ttf") format('truetype');
-// }
-
-// @font-face {
-//     font-family: 'Exo 2';
-//     font-style: normal;
-//     font-weight: 800;
-//     src: url("~assets/fonts/Exo_2/static/Exo2-ExtraBold.ttf") format('truetype');
-// }
