+++ layouts/default.vue
@@ -1,19 +1,21 @@
 <template>
-    <div>
-        <BaseHeader />
-        <Nuxt />
-    </div>
+  <div>
+    <!-- Header Menu for Desktop Devices -->
+    <BaseHeader />
+    <!-- Main page Content -->
+    <Nuxt />
+    <!-- Footer Menu for Mobile Devices -->
+  </div>
 </template>
 
 <script>
 import BaseHeader from "/components/BaseHeader.vue";
 export default {
-    components: {
-        BaseHeader
-    }
-}
+  components: {
+    BaseHeader,
+  },
+};
 </script>
 
 <style>
-
 </style>
