+++ pages/analytics/side-menu.vue
@@ -92,7 +92,7 @@
 </template>
 
 <script>
-import { getServerCustomDateString, getDateInputValue } from "../utils";
+import { getServerCustomDateString, getDateInputValue } from "../../utils";
 import { mapGetters } from "vuex";
 export default {
   computed: {
@@ -217,219 +217,3 @@ export default {
   },
 };
 </script>
-
-<style lang="scss">
-/* This needs to be global */
-.mj-daterange-picker .preset-ranges .preset {
-  font-size: 0.875rem;
-}
-.custom-control-input:checked ~ .custom-control-label::before {
-  border-color: #fd624f;
-  background-color: #fd624f;
-  box-shadow: none;
-}
-
-.custom-control-input:focus ~ .custom-control-label::before {
-  box-shadow: 0 0 0 0.2rem rgba(253, 98, 79, 0.25);
-}
-
-.custom-control-input:focus:not(:checked) ~ .custom-control-label::before {
-  border-color: rgba(253, 98, 79, 0.25);
-}
-
-
-.divider {
-  margin-left: 1.25rem;
-}
-
-.side-menu {
-  padding-right: 1.25rem;
-
-  .filters-scroll {
-    max-height: calc(100vh - 14.5rem);
-    overflow-y: scroll;
-    overflow-x: hidden;
-
-    /* Hide scrollbar for IE, Edge and Firefox */
-    -ms-overflow-style: none; /* IE and Edge */
-    scrollbar-width: none; /* Firefox */
-
-    /* Hide scrollbar for Chrome, Safari and Opera */
-    &::-webkit-scrollbar {
-      display: none;
-    }
-  }
-
-  .field-title {
-    padding-left: 1.25rem;
-  }
-}
-.hourly-checkbox{
-    margin: 0.2rem 0;
-    font-family: ProductSans;
-    font-size: 0.875rem;
-    line-height: 1.5rem;
-    color: #202124;
-    span{
-        font-size: 0.5rem;
-    }
-}
-.side-menu-item {
-  .side-menu-button {
-    /* font-family: ProductSans; */
-    font-family: ProductSans;
-    font-size: 0.875rem;
-    line-height: 2.5rem;
-    color: #202124;
-    border-radius: 0 0.375rem 0.375rem 0;
-    background-color: transparent;
-    border: none;
-    display: block;
-    min-width: 100%;
-    text-align: left;
-    padding-left: 1.25rem;
-    cursor: pointer;
-    height: 2.625rem;
-
-    &:hover {
-      // color: #fd624f;
-      background-color: #F3f3f3;
-      border: none;
-    }
-    &:active {
-      background-color: #F3f3f3;
-      border: none;
-      color: #000b12;
-      border: none;
-    }
-    &:focus, &:active {
-      box-shadow: none !important;
-    }
-  }
-  .childs-list {
-    margin: 0 !important;
-
-    .side-menu-button {
-      padding-left: 1.875rem;
-    }
-  }
-  .side-menu-button-active {
-    color: #fd624f;
-    // background-color: #F3f3f3;
-    border-radius: 0 0.375rem 0.375rem 0;
-    border: none;
-    &:focus {
-      box-shadow: none;
-    }
-    &.not-collapsed {
-      color: #000b12;
-    }
-  }
-
-  .side-menu-item {
-    font-family: ProductSans;
-    font-size: 0.875rem;
-    line-height: 2.5rem;
-    color: #202124;
-    // height: 42px;
-    background-color: transparent;
-    border: none;
-    display: block;
-    min-width: 100%;
-    text-align: left;
-    padding: 0 1.25rem;
-    border-radius: 0 0.375rem 0.375rem 0;
-    cursor: pointer;
-    box-shadow: none !important;
-
-    &.not-collapsed {
-      img {
-        transform: rotate(180deg);
-      }
-    }
-
-    &:hover {
-      // color: #fd624f;
-      background-color: #F3f3f3;
-      border-radius: 0 0.375rem 0.375rem 0;
-      border: none;
-    }
-  }
-  .side-menu-item.side-menu-button-active {
-    color: #fd624f;
-    // background-color: #F3f3f3;
-    border-radius: 0 0.375rem 0.375rem 0;
-    border: none;
-  }
-  .side-menu-item.not-collapsed {
-    color: #000b12;
-    background-color: #F3f3f3;
-    border-radius: 0 0.375rem 0.375rem 0;
-    border: none;
-  }
-  .side-menu-item.collapsed {
-    color: #000b12;
-  }
-  .side-menu-single-active {
-    color: #fd624f !important;
-    border-radius: 0 0.375rem 0.375rem 0;
-    border: none;
-  }
-
-  .childs-list {
-    margin-left: 1rem;
-  }
-}
-.date-range-field {
-  position: relative;
-  .date-picker-holder {
-    z-index: 1000;
-    background-color: #fff;
-    position: absolute;
-    transform: scale(0.8);
-    transform-origin: left top;
-    left: 100%;
-    top: 0;
-  }
-
-  .date-input {
-    display: flex;
-    justify-content: space-between;
-    font-family: ProductSans;
-    line-height: 2.5rem;
-    color: #202124;
-    padding: 0 1.25rem;
-    user-select: none;
-    &:hover {
-      background-color: #F3f3f3;
-      border: none;
-      color: #000b12;
-      border-radius: 0 0.375rem 0.375rem 0;
-      border: none;
-      cursor: pointer;
-    }
-  }
-  .date-input-active {
-    background-color: #F3f3f3;
-    border: none;
-    color: #000b12;
-    border-radius: 0;
-    border: none;
-    img {
-      transform: rotate(270deg);
-    }
-  }
-}
-
-.mj-daterange-picker {
-}
-
-.field-title {
-  font-family: ProductSans;
-  font-size: 0.875rem;
-  line-height: 2.5rem;
-  color: #5f6368;
-}
-
-
-</style>
