+++ plugins/vue-handsontable.js
@@ -1,4 +1,5 @@
 import Vue from 'vue';
+import Handsontable from 'handsontable';
 const HotTable = !process.client ? null : require('@handsontable/vue').HotTable
 const HotColumn = !process.client ? null : require('@handsontable/vue').HotColumn
 
@@ -9,3 +10,7 @@ Vue.mixin({
     HotColumn
   }
 })
+
+export default {
+  Handsontable
+}
